refactor(consultation): 운영자 상담 realtime 상태를 분리
